pyproject.toml
[project]
name = "dm-streamvalve"
version = "0.0.4"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Bastien Chevreux", email = "bach@chevreux.org" }
]
requires-python = ">=3.13"
dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ollama>=0.4.7",
    "pre-commit>=4.1.0",
    "pytest-cov>=6.0.0",
    "rich>=13.9.4",
]
